Added support for a margin around a tileset image
The margin attribute has been added to the tileset element, next to the
tile spacing attribute. This basically implements Flyspray task 79, which was
about adding support for an offset. The margin also applies to the bottom and
right sides, however.

+++ src/tiled/core/MapObject.java
@@ -22,7 +22,7 @@ public class MapObject implements Cloneable
 {
     private Properties properties = new Properties();
 
-    protected float x, y;
+    protected int x, y;
     protected Rectangle bounds = new Rectangle();
     protected boolean bVisible = true;
     protected String name = "Object";
@@ -82,11 +82,11 @@ public class MapObject implements Cloneable
     }
 
     public int getX() {
-        return (int) x;
+        return x;
     }
 
     public int getY() {
-        return (int) y;
+        return y;
     }
 
     public String getName() {
